+++ src/Console/Command/ConfigGeneratorCommand.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Crunz\Console\Command;
 
-use Crunz\Output\VerbosityAwareOutput;
+use Crunz\Filesystem\FilesystemInterface;
+use Crunz\Path\Path;
 use Crunz\Timezone\ProviderInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -17,11 +18,17 @@ class ConfigGeneratorCommand extends Command
     /** @var ProviderInterface */
     private $timezoneProvider;
     /** @var Filesystem */
+    private $symfonyFilesystem;
+    /** @var FilesystemInterface */
     private $filesystem;
 
-    public function __construct(ProviderInterface $timezoneProvider, Filesystem $filesystem)
-    {
+    public function __construct(
+        ProviderInterface $timezoneProvider,
+        Filesystem $symfonyFilesystem,
+        FilesystemInterface $filesystem
+    ) {
         $this->timezoneProvider = $timezoneProvider;
+        $this->symfonyFilesystem = $symfonyFilesystem;
         $this->filesystem = $filesystem;
 
         parent::__construct();
@@ -44,22 +51,22 @@ class ConfigGeneratorCommand extends Command
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $verbosityAwareOutput = new VerbosityAwareOutput($output);
         $symfonyStyleIo = new SymfonyStyle($input, $output);
-
-        $path = getbase() . '/crunz.yml';
+        $cwd = $this->filesystem
+            ->getCwd();
+        $path = Path::create([$cwd, 'crunz.yml'])->toString();
         $destination = \realpath($path) ?: $path;
         $configExists = $this->filesystem
-            ->exists($destination)
+            ->fileExists($destination)
         ;
 
-        $verbosityAwareOutput->writeln(
+        $output->writeln(
             "<info>Destination config file: '{$destination}'.</info>",
             OutputInterface::VERBOSITY_VERBOSE
         );
 
         if ($configExists) {
-            $verbosityAwareOutput->writeln(
+            $output->writeln(
                 "<comment>The configuration file already exists at '{$destination}'.</comment>"
             );
 
@@ -67,12 +74,12 @@ class ConfigGeneratorCommand extends Command
         }
 
         $src = __DIR__ . '/../../../crunz.yml';
-        $verbosityAwareOutput->writeln(
+        $output->writeln(
             "<info>Source config file: '{$src}'.</info>",
             OutputInterface::VERBOSITY_VERBOSE
         );
         $defaultTimezone = $this->askForTimezone($symfonyStyleIo);
-        $verbosityAwareOutput->writeln(
+        $output->writeln(
             "<info>Provided timezone: '{$defaultTimezone}'.</info>",
             OutputInterface::VERBOSITY_VERBOSE
         );
@@ -122,7 +129,7 @@ class ConfigGeneratorCommand extends Command
         $src,
         $timezone
     ) {
-        $this->filesystem
+        $this->symfonyFilesystem
             ->dumpFile(
                 $destination,
                 \str_replace(