+++ src/Console/Command/ScheduleRunCommand.php
@@ -8,9 +8,11 @@ use Crunz\Configuration\Configuration;
 use Crunz\EventRunner;
 use Crunz\Schedule;
 use Crunz\Task\Collection;
+use Crunz\Task\TaskNumber;
 use Crunz\Task\Timezone;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class ScheduleRunCommand extends Command
@@ -29,17 +31,21 @@ class ScheduleRunCommand extends Command
     private $eventRunner;
     /** @var Timezone */
     private $taskTimezone;
+    /** @var Schedule\ScheduleFactory */
+    private $scheduleFactory;
 
     public function __construct(
         Collection $taskCollection,
         Configuration $configuration,
         EventRunner $eventRunner,
-        Timezone $taskTimezone
+        Timezone $taskTimezone,
+        Schedule\ScheduleFactory $scheduleFactory
     ) {
         $this->taskCollection = $taskCollection;
         $this->configuration = $configuration;
         $this->eventRunner = $eventRunner;
         $this->taskTimezone = $taskTimezone;
+        $this->scheduleFactory = $scheduleFactory;
 
         parent::__construct();
     }
@@ -49,10 +55,6 @@ class ScheduleRunCommand extends Command
      */
     protected function configure()
     {
-        $sourcePath = $this->configuration
-            ->get('source')
-        ;
-
         $this->setName('schedule:run')
             ->setDescription('Starts the event runner.')
             ->setDefinition(
@@ -61,9 +63,23 @@ class ScheduleRunCommand extends Command
                         'source',
                         InputArgument::OPTIONAL,
                         'The source directory for collecting the task files.',
-                        generate_path($sourcePath)
+                        $this->configuration
+                            ->getSourcePath()
                     ),
                 ]
+            )
+            ->addOption(
+                'force',
+                'f',
+                InputOption::VALUE_NONE,
+                'Run all tasks regardless of configured run time.'
+            )
+            ->addOption(
+                'task',
+                't',
+                InputOption::VALUE_REQUIRED,
+                'Which task to run. Provide task number from <info>schedule:list</info> command.',
+                null
             )
            ->setHelp('This command starts the Crunz event runner.');
     }
@@ -75,6 +91,7 @@ class ScheduleRunCommand extends Command
     {
         $this->arguments = $input->getArguments();
         $this->options = $input->getOptions();
+        $task = $this->options['task'];
         $files = $this->taskCollection
             ->all($this->arguments['source'])
         ;
@@ -87,7 +104,6 @@ class ScheduleRunCommand extends Command
 
         // List of schedules
         $schedules = [];
-
         $tasksTimezone = $this->taskTimezone
             ->timezoneForComparisons()
         ;
@@ -98,17 +114,38 @@ class ScheduleRunCommand extends Command
                 continue;
             }
 
-            // We keep the events which are due and dismiss the rest.
-            $schedule->events(
-                $schedule->dueEvents(
-                    $tasksTimezone
-                )
-            );
-
             if (\count($schedule->events())) {
                 $schedules[] = $schedule;
             }
         }
+
+        // Is specified task should be invoked?
+        if (null !== $task) {
+            $schedules = $this->scheduleFactory
+                ->singleTaskSchedule(TaskNumber::fromString($task), ...$schedules);
+        }
+
+        $schedules = \array_map(
+            function (Schedule $schedule) use ($tasksTimezone) {
+                if (false === $this->options['force']) {
+                    // We keep the events which are due and dismiss the rest.
+                    $schedule->events(
+                        $schedule->dueEvents(
+                            $tasksTimezone
+                        )
+                    );
+                }
+
+                return $schedule;
+            },
+            $schedules
+        );
+        $schedules = \array_filter(
+            $schedules,
+            function (Schedule $schedule) {
+                return \count($schedule->events());
+            }
+        );
 
         if (!\count($schedules)) {
             $output->writeln('<comment>No event is due!</comment>');
@@ -120,5 +157,7 @@ class ScheduleRunCommand extends Command
         $this->eventRunner
             ->handle($output, $schedules)
         ;
+
+        return 0;
     }
 }